feat: Add multi-language support, add Traditional Chinese translation files and update existing translations
