+++ webview-ui/src/i18n/zgsm_locales/en/chat.json
@@ -1,46 +1,46 @@
 {
-	"greeting": "What can Zhuge Shenma do for you?",
+	"greeting": "What can Shenma do for you?",
 	"aboutMe": "Thanks to the latest breakthroughs in agentic coding capabilities, I can handle complex software development tasks step-by-step. With tools that let me create & edit files, explore complex projects, use the browser, and execute terminal commands (after you grant permission), I can assist you in ways that go beyond code completion or tech support. I can even use MCP to create new tools and extend my own capabilities.",
 	"instructions": {
-		"wantsToFetch": "Zhuge Shenma wants to fetch detailed instructions to assist with the current task"
+		"wantsToFetch": "Shenma wants to fetch detailed instructions to assist with the current task"
 	},
 	"fileOperations": {
-		"wantsToRead": "Zhuge Shenma wants to read this file:",
-		"wantsToReadOutsideWorkspace": "Zhuge Shenma wants to read this file outside of the workspace:",
-		"didRead": "Zhuge Shenma read this file:",
-		"wantsToEdit": "Zhuge Shenma wants to edit this file:",
-		"wantsToEditOutsideWorkspace": "Zhuge Shenma wants to edit this file outside of the workspace:",
-		"wantsToCreate": "Zhuge Shenma wants to create a new file:"
+		"wantsToRead": "Shenma wants to read this file:",
+		"wantsToReadOutsideWorkspace": "Shenma wants to read this file outside of the workspace:",
+		"didRead": "Shenma read this file:",
+		"wantsToEdit": "Shenma wants to edit this file:",
+		"wantsToEditOutsideWorkspace": "Shenma wants to edit this file outside of the workspace:",
+		"wantsToCreate": "Shenma wants to create a new file:"
 	},
 	"directoryOperations": {
-		"wantsToViewTopLevel": "Zhuge Shenma wants to view the top level files in this directory:",
-		"didViewTopLevel": "Zhuge Shenma viewed the top level files in this directory:",
-		"wantsToViewRecursive": "Zhuge Shenma wants to recursively view all files in this directory:",
-		"didViewRecursive": "Zhuge Shenma recursively viewed all files in this directory:",
-		"wantsToViewDefinitions": "Zhuge Shenma wants to view source code definition names used in this directory:",
-		"didViewDefinitions": "Zhuge Shenma viewed source code definition names used in this directory:",
-		"wantsToSearch": "Zhuge Shenma wants to search this directory for <code>{{regex}}</code>:",
-		"didSearch": "Zhuge Shenma searched this directory for <code>{{regex}}</code>:"
+		"wantsToViewTopLevel": "Shenma wants to view the top level files in this directory:",
+		"didViewTopLevel": "Shenma viewed the top level files in this directory:",
+		"wantsToViewRecursive": "Shenma wants to recursively view all files in this directory:",
+		"didViewRecursive": "Shenma recursively viewed all files in this directory:",
+		"wantsToViewDefinitions": "Shenma wants to view source code definition names used in this directory:",
+		"didViewDefinitions": "Shenma viewed source code definition names used in this directory:",
+		"wantsToSearch": "Shenma wants to search this directory for <code>{{regex}}</code>:",
+		"didSearch": "Shenma searched this directory for <code>{{regex}}</code>:"
 	},
 	"mcp": {
-		"wantsToUseTool": "Zhuge Shenma wants to use a tool on the {{serverName}} MCP server:",
-		"wantsToAccessResource": "Zhuge Shenma wants to access a resource on the {{serverName}} MCP server:"
+		"wantsToUseTool": "Shenma wants to use a tool on the {{serverName}} MCP server:",
+		"wantsToAccessResource": "Shenma wants to access a resource on the {{serverName}} MCP server:"
 	},
 	"modes": {
-		"wantsToSwitch": "Zhuge Shenma wants to switch to {{mode}} mode",
-		"wantsToSwitchWithReason": "Zhuge Shenma wants to switch to {{mode}} mode because: {{reason}}",
-		"didSwitch": "Zhuge Shenma switched to {{mode}} mode",
-		"didSwitchWithReason": "Zhuge Shenma switched to {{mode}} mode because: {{reason}}"
+		"wantsToSwitch": "Shenma wants to switch to {{mode}} mode",
+		"wantsToSwitchWithReason": "Shenma wants to switch to {{mode}} mode because: {{reason}}",
+		"didSwitch": "Shenma switched to {{mode}} mode",
+		"didSwitchWithReason": "Shenma switched to {{mode}} mode because: {{reason}}"
 	},
 	"subtasks": {
-		"wantsToCreate": "Zhuge Shenma wants to create a new subtask in {{mode}} mode:",
-		"wantsToFinish": "Zhuge Shenma wants to finish this subtask"
+		"wantsToCreate": "Shenma wants to create a new subtask in {{mode}} mode:",
+		"wantsToFinish": "Shenma wants to finish this subtask"
 	},
 	"questions": {
-		"hasQuestion": "Zhuge Shenma has a question:"
+		"hasQuestion": "Shenma has a question:"
 	},
-	"troubleMessage": "Zhuge Shenma is having trouble...",
+	"troubleMessage": "Shenma is having trouble...",
 	"browser": {
-		"rooWantsToUse": "Zhuge Shenma wants to use the browser:"
+		"rooWantsToUse": "Shenma wants to use the browser:"
 	}
 }
