+++ webview-ui/src/i18n/zgsm_locales/en/mcp.json
@@ -1,9 +1,9 @@
 {
-	"description": "The <0>Model Context Protocol</0> enables communication with locally running MCP servers that provide additional tools and resources to extend Zhuge Shenma's capabilities. You can use <1>community-made servers</1> or ask Zhuge Shenma to create new tools specific to your workflow (e.g., \"add a tool that gets the latest npm docs\").",
+	"description": "The <0>Model Context Protocol</0> enables communication with locally running MCP servers that provide additional tools and resources to extend Shenma's capabilities. You can use <1>community-made servers</1> or ask Shenma to create new tools specific to your workflow (e.g., \"add a tool that gets the latest npm docs\").",
 	"enableToggle": {
-		"description": "When enabled, Zhuge Shenma will be able to interact with MCP servers for advanced functionality. If you're not using MCP, you can disable this to reduce Zhuge Shenma's token usage."
+		"description": "When enabled, Shenma will be able to interact with MCP servers for advanced functionality. If you're not using MCP, you can disable this to reduce Shenma's token usage."
 	},
 	"enableServerCreation": {
-		"description": "When enabled, Zhuge Shenma can help you create new MCP servers via commands like \"add a new tool to...\". If you don't need to create MCP servers you can disable this to reduce Zhuge Shenma's token usage."
+		"description": "When enabled, Shenma can help you create new MCP servers via commands like \"add a new tool to...\". If you don't need to create MCP servers you can disable this to reduce Shenma's token usage."
 	}
 }
