+++ webview-ui/src/i18n/zgsm_locales/en/prompts.json
@@ -1,23 +1,23 @@
 {
 	"modes": {
-		"createModeHelpText": "Hit the + to create a new custom mode, or just ask Zhuge Shenma in chat to create one for you!"
+		"createModeHelpText": "Hit the + to create a new custom mode, or just ask Shenma in chat to create one for you!"
 	},
 	"roleDefinition": {
-		"description": "Define Zhuge Shenma's expertise and personality for this mode. This description shapes how Zhuge Shenma presents itself and approaches tasks."
+		"description": "Define Shenma's expertise and personality for this mode. This description shapes how Shenma presents itself and approaches tasks."
 	},
 	"globalCustomInstructions": {
-		"description": "These instructions apply to all modes. They provide a base set of behaviors that can be enhanced by mode-specific instructions below.\nIf you would like Zhuge Shenma to think and speak in a different language than your editor display language ({{language}}), you can specify it here."
+		"description": "These instructions apply to all modes. They provide a base set of behaviors that can be enhanced by mode-specific instructions below.\nIf you would like Shenma to think and speak in a different language than your editor display language ({{language}}), you can specify it here."
 	},
 	"supportPrompts": {
 		"types": {
 			"ENHANCE": {
-				"description": "Use prompt enhancement to get tailored suggestions or improvements for your inputs. This ensures Zhuge Shenma understands your intent and provides the best possible responses. Available via the ✨ icon in chat."
+				"description": "Use prompt enhancement to get tailored suggestions or improvements for your inputs. This ensures Shenma understands your intent and provides the best possible responses. Available via the ✨ icon in chat."
 			}
 		}
 	},
 	"createModeDialog": {
 		"roleDefinition": {
-			"description": "Define Zhuge Shenma's expertise and personality for this mode."
+			"description": "Define Shenma's expertise and personality for this mode."
 		}
 	}
 }
