+++ webview-ui/src/i18n/zgsm_locales/en/welcome.json
@@ -1,6 +1,6 @@
 {
-	"greeting": "Hi, I'm Zhuge Shenma!",
+	"greeting": "Hi, I'm Shenma!",
 	"telemetry": {
-		"title": "Help Improve Zhuge Shenma"
+		"title": "Help Improve Shenma"
 	}
 }
