+++ webview-ui/src/i18n/zgsm_locales/zh-TW/chat.json
@@ -1,0 +1,46 @@
+{
+	"greeting": "諸葛神碼可以為您做些什麼？",
+	"aboutMe": "得益於最新的代理編碼能力突破，我可以逐步處理複雜的軟件開發任務。通過允許我創建和編輯文件、探索複雜項目、使用瀏覽器以及執行終端命令（在您授予權限後）的工具，我可以以超越代碼補全或技術支持的方式協助您。我甚至可以使用MCP創建新工具並擴展自己的能力。",
+	"instructions": {
+		"wantsToFetch": "諸葛神碼想要獲取詳細指令以協助當前任務"
+	},
+	"fileOperations": {
+		"wantsToRead": "諸葛神碼想要讀取此文件：",
+		"wantsToReadOutsideWorkspace": "諸葛神碼想要讀取工作區外的此文件：",
+		"didRead": "諸葛神碼已讀取此文件：",
+		"wantsToEdit": "諸葛神碼想要編輯此文件：",
+		"wantsToEditOutsideWorkspace": "諸葛神碼想要編輯工作區外的此文件：",
+		"wantsToCreate": "諸葛神碼想要創建新文件："
+	},
+	"directoryOperations": {
+		"wantsToViewTopLevel": "諸葛神碼想要查看此目錄中的頂層文件：",
+		"didViewTopLevel": "諸葛神碼已查看此目錄中的頂層文件：",
+		"wantsToViewRecursive": "諸葛神碼想要遞歸查看此目錄中的所有文件：",
+		"didViewRecursive": "諸葛神碼已遞歸查看此目錄中的所有文件：",
+		"wantsToViewDefinitions": "諸葛神碼想要查看此目錄中使用的源代碼定義名稱：",
+		"didViewDefinitions": "諸葛神碼已查看此目錄中使用的源代碼定義名稱：",
+		"wantsToSearch": "諸葛神碼想要在此目錄中搜索<code>{{regex}}</code>：",
+		"didSearch": "諸葛神碼已在此目錄中搜索<code>{{regex}}</code>："
+	},
+	"mcp": {
+		"wantsToUseTool": "諸葛神碼想要在{{serverName}} MCP服務器上使用工具：",
+		"wantsToAccessResource": "諸葛神碼想要訪問{{serverName}} MCP服務器上的資源："
+	},
+	"modes": {
+		"wantsToSwitch": "諸葛神碼想要切換到{{mode}}模式",
+		"wantsToSwitchWithReason": "諸葛神碼想要切換到{{mode}}模式，因為：{{reason}}",
+		"didSwitch": "諸葛神碼已切換到{{mode}}模式",
+		"didSwitchWithReason": "諸葛神碼已切換到{{mode}}模式，因為：{{reason}}"
+	},
+	"subtasks": {
+		"wantsToCreate": "諸葛神碼想要在{{mode}}模式下創建新子任務：",
+		"wantsToFinish": "諸葛神碼想要完成此子任務"
+	},
+	"questions": {
+		"hasQuestion": "諸葛神碼有一個問題："
+	},
+	"troubleMessage": "諸葛神碼遇到了一些問題...",
+	"browser": {
+		"rooWantsToUse": "諸葛神碼想要使用瀏覽器："
+	}
+}
