+++ webview-ui/src/i18n/zgsm_locales/zh-TW/prompts.json
@@ -1,0 +1,23 @@
+{
+	"modes": {
+		"createModeHelpText": "點選 + 建立新的自訂模式，或者在聊天中直接請諸葛神碼為您建立！"
+	},
+	"roleDefinition": {
+		"description": "定義此模式下諸葛神碼的專業知識和個性。此描述會形塑諸葛神碼如何展現自己並處理工作。"
+	},
+	"globalCustomInstructions": {
+		"description": "這些指令適用於所有模式。它們提供了一組基本行為，可以透過下方的模式專屬自訂指令來強化。\n如果您希望諸葛神碼使用與編輯器顯示語言 ({{language}}) 不同的語言來思考和對話，您可以在這裡指定。"
+	},
+	"supportPrompts": {
+		"types": {
+			"ENHANCE": {
+				"description": "使用提示詞增強功能取得針對您輸入的客製化建議或改進。這確保 諸葛神碼 能理解您的意圖並提供最佳的回應。可透過聊天中的 ✨ 圖示使用。"
+			}
+		}
+	},
+	"createModeDialog": {
+		"roleDefinition": {
+			"description": "定義諸葛神碼的專業領域和個性。"
+		}
+	}
+}
