+++ webview-ui/src/i18n/zgsm_locales/zh-TW/welcome.json
@@ -1,0 +1,6 @@
+{
+	"greeting": "你好，我是諸葛神碼！",
+	"telemetry": {
+		"title": "幫助改進諸葛神碼"
+	}
+}
