+++ zgsm/src/i18n/locales/en/common.json
@@ -1,12 +1,12 @@
 {
 	"extension": {
-		"name": "Zhuge Shenma"
+		"name": "Shenma"
 	},
 	"storage": {
 		"path_placeholder": "D:\\RooCodeStorage",
 		"enter_absolute_path": "Please enter an absolute path (e.g. D:\\RooCodeStorage or /home/user/storage)"
 	},
 	"input": {
-		"task_prompt": "What should Zhuge Shenma do?"
+		"task_prompt": "What should Shenma do?"
 	}
 }
